requirements.txt
aniso8601==9.0.1
attrs==22.2.0
autopep8==2.0.2
bcrypt==4.0.1
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0
exceptiongroup==1.1.1
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.4
flask-marshmallow==0.15.0
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
flask-restx==1.1.0
Flask-Testing==0.8.1
Flask-WTF==1.1.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
logconfig==0.4.0
logutils==0.3.5
MarkupSafe==2.1.2
marshmallow==3.19.0
mongoengine==0.27.0
numpy==1.24.2
packaging==23.1
pandas==2.0.0
pluggy==1.0.0
pycodestyle==2.10.0
PyJWT==2.6.0
pymongo==4.3.3
pyrsistent==0.19.3
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
tomli==2.0.1
tzdata==2023.3
Werkzeug==2.2.3
WTForms==3.0.1
xgboost==1.7.5
